+++ relayer-cli/src/commands/tx/transfer.rs
@@ -1,18 +1,17 @@
-use std::sync::Arc;
-
 use abscissa_core::{config::Override, Command, FrameworkErrorKind, Options, Runnable};
 
-use tokio::runtime::Runtime as TokioRuntime;
-
-use ibc::events::IbcEvent;
-use ibc::ics02_client::height::Height;
-use ibc::ics24_host::identifier::{ChainId, ChannelId, PortId};
+use ibc::{
+    events::IbcEvent,
+    ics02_client::client_state::ClientState,
+    ics02_client::height::Height,
+    ics24_host::identifier::{ChainId, ChannelId, PortId},
+};
 use ibc_relayer::{
-    chain::{Chain, CosmosSdkChain},
     config::Config,
     transfer::{build_and_send_transfer_messages, TransferOptions},
 };
 
+use crate::cli_utils::ChainHandlePair;
 use crate::conclude::{exit_with_unrecoverable_error, Output};
 use crate::error::{Error, Kind};
 use crate::prelude::*;
@@ -135,48 +134,33 @@ impl Runnable for TxIcs20MsgTransferCmd {
 
         debug!("Message: {:?}", opts);
 
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-
-        let src_chain_res =
-            CosmosSdkChain::bootstrap(opts.packet_src_chain_config.clone(), rt.clone())
-                .map_err(|e| Kind::Runtime.context(e));
-
-        let src_chain = match src_chain_res {
-            Ok(chain) => chain,
-            Err(e) => return Output::error(format!("{}", e)).exit(),
-        };
-
-        let dst_chain_res = CosmosSdkChain::bootstrap(opts.packet_dst_chain_config.clone(), rt)
-            .map_err(|e| Kind::Runtime.context(e));
-
-        let dst_chain = match dst_chain_res {
-            Ok(chain) => chain,
-            Err(e) => return Output::error(format!("{}", e)).exit(),
-        };
+        let chains = ChainHandlePair::spawn(&config, &self.src_chain_id, &self.dst_chain_id)
+            .unwrap_or_else(exit_with_unrecoverable_error);
 
         // Double check that channels and chain identifiers match.
         // To do this, fetch from the source chain the channel end, then the associated connection
         // end, and then the underlying client state; finally, check that this client is verifying
         // headers for the destination chain.
-        let channel_end = src_chain
+        let channel_end_src = chains
+            .src
             .query_channel(
                 &opts.packet_src_port_id,
                 &opts.packet_src_channel_id,
                 Height::zero(),
             )
             .unwrap_or_else(exit_with_unrecoverable_error);
-        if !channel_end.is_open() {
+        if !channel_end_src.is_open() {
             return Output::error(format!(
                 "the requested port/channel ('{}'/'{}') on chain id '{}' is in state '{}'; expected 'open' state",
                 opts.packet_src_port_id,
                 opts.packet_src_channel_id,
                 self.src_chain_id,
-                channel_end.state
+                channel_end_src.state
             ))
             .exit();
         }
 
-        let conn_id = match channel_end.connection_hops.first() {
+        let conn_id = match channel_end_src.connection_hops.first() {
             None => {
                 return Output::error(format!(
                     "could not retrieve the connection hop underlying port/channel '{}'/'{}' on chain '{}'",
@@ -187,13 +171,15 @@ impl Runnable for TxIcs20MsgTransferCmd {
             Some(cid) => cid,
         };
 
-        let conn_end = src_chain
+        let conn_end = chains
+            .src
             .query_connection(conn_id, Height::zero())
             .unwrap_or_else(exit_with_unrecoverable_error);
 
         debug!("connection hop underlying the channel: {:?}", conn_end);
 
-        let src_chain_client_state = src_chain
+        let src_chain_client_state = chains
+            .src
             .query_client_state(conn_end.client_id(), Height::zero())
             .unwrap_or_else(exit_with_unrecoverable_error);
 
@@ -202,18 +188,18 @@ impl Runnable for TxIcs20MsgTransferCmd {
             src_chain_client_state
         );
 
-        if src_chain_client_state.chain_id != self.dst_chain_id {
+        if src_chain_client_state.chain_id() != self.dst_chain_id {
             return Output::error(
                 format!("the requested port/channel ('{}'/'{}') provides a path from chain '{}' to \
                  chain '{}' (not to the destination chain '{}'). Bailing due to mismatching arguments.",
                         opts.packet_src_port_id, opts.packet_src_channel_id,
                         self.src_chain_id,
-                        src_chain_client_state.chain_id, self.dst_chain_id)).exit();
+                        src_chain_client_state.chain_id(), self.dst_chain_id)).exit();
         }
 
         // Checks pass, build and send the tx
         let res: Result<Vec<IbcEvent>, Error> =
-            build_and_send_transfer_messages(src_chain, dst_chain, opts)
+            build_and_send_transfer_messages(chains.src, chains.dst, opts)
                 .map_err(|e| Kind::Tx.context(e).into());
 
         match res {