+++ relayer/src/channel.rs
@@ -4,7 +4,7 @@ use prost_types::Any;
 use serde::Serialize;
 use std::time::Duration;
 use thiserror::Error;
-use tracing::{debug, error};
+use tracing::{debug, error, info};
 
 use ibc::events::IbcEvent;
 use ibc::ics04_channel::channel::{ChannelEnd, Counterparty, IdentifiedChannelEnd, Order, State};
@@ -14,22 +14,39 @@ use ibc::ics04_channel::msgs::chan_open_ack::MsgChannelOpenAck;
 use ibc::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
 use ibc::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
 use ibc::ics04_channel::msgs::chan_open_try::MsgChannelOpenTry;
-use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
+use ibc::ics24_host::identifier::{
+    ChainId, ChannelId, ClientId, ConnectionId, PortChannelId, PortId,
+};
 use ibc::tx_msg::Msg;
 use ibc::Height;
+use ibc_proto::ibc::core::channel::v1::QueryConnectionChannelsRequest;
 
+use crate::chain::counterparty::{channel_connection_client, channel_state_on_destination};
 use crate::chain::handle::ChainHandle;
 use crate::connection::Connection;
 use crate::error::Error;
 use crate::foreign_client::{ForeignClient, ForeignClientError};
 use crate::object::Channel as WorkerChannelObject;
 use crate::supervisor::Error as WorkerChannelError;
-
-const MAX_RETRIES: usize = 5;
-
-use crate::chain::counterparty::{channel_connection_client, channel_state_on_destination};
 use crate::util::retry::RetryResult;
-use ibc_proto::ibc::core::channel::v1::QueryConnectionChannelsRequest;
+use crate::util::retry::{retry_count, retry_with_index};
+
+mod retry_strategy {
+    use std::time::Duration;
+
+    use retry::delay::Fibonacci;
+
+    use crate::util::retry::clamp_total;
+
+    // Default parameters for the retrying mechanism
+    const MAX_DELAY: Duration = Duration::from_secs(60); // 1 minute
+    const MAX_TOTAL_DELAY: Duration = Duration::from_secs(10 * 60); // 10 minutes
+    const INITIAL_DELAY: Duration = Duration::from_secs(1); // 1 second
+
+    pub fn default() -> impl Iterator<Item = Duration> {
+        clamp_total(Fibonacci::from(INITIAL_DELAY), MAX_DELAY, MAX_TOTAL_DELAY)
+    }
+}
 
 #[derive(Debug, Error)]
 pub enum ChannelError {
@@ -55,6 +72,15 @@ pub enum ChannelError {
         "failed during a transaction submission step to chain id {0} with underlying error: {1}"
     )]
     SubmitError(ChainId, Error),
+
+    #[error("the channel is partially open ({0}, {1})")]
+    PartialOpenHandshake(State, State),
+
+    #[error("channel {0} on chain {1} has no counterparty channel id")]
+    IncompleteChannelState(PortChannelId, ChainId),
+
+    #[error("channel {0} on chain {1} expected to have counterparty {2} (but instead has {3})")]
+    MismatchingChannelEnds(PortChannelId, ChainId, PortChannelId, PortChannelId),
 }
 
 #[derive(Clone, Debug, Serialize)]
@@ -328,125 +354,218 @@ impl Channel {
         }
     }
 
+    fn do_chan_open_init_and_send(&mut self) -> Result<(), ChannelError> {
+        let event = self
+            .flipped()
+            .build_chan_open_init_and_send()
+            .map_err(|e| {
+                error!("Failed ChanInit {:?}: {:?}", self.a_side, e);
+                e
+            })?;
+
+        info!("done {} => {:#?}\n", self.src_chain().id(), event);
+
+        let channel_id = extract_channel_id(&event)?;
+        self.a_side.channel_id = Some(channel_id.clone());
+        info!("successfully opened init channel");
+
+        Ok(())
+    }
+
+    // Check that the channel was created on a_chain
+    fn do_chan_open_init_and_send_with_retry(&mut self) -> Result<(), ChannelError> {
+        retry_with_index(retry_strategy::default(), |_| {
+            self.do_chan_open_init_and_send()
+        })
+        .map_err(|err| {
+            error!("failed to open channel after {} retries", err);
+            ChannelError::Failed(format!(
+                "Failed to finish channel open init in {} iterations for {:?}",
+                retry_count(&err),
+                self
+            ))
+        })?;
+
+        Ok(())
+    }
+
+    fn do_chan_open_try_and_send(&mut self) -> Result<(), ChannelError> {
+        let event = self.build_chan_open_try_and_send().map_err(|e| {
+            error!("Failed ChanTry {:?}: {:?}", self.b_side, e);
+            e
+        })?;
+
+        let channel_id = extract_channel_id(&event)?;
+        self.b_side.channel_id = Some(channel_id.clone());
+
+        println!("done {} => {:#?}\n", self.dst_chain().id(), event);
+        Ok(())
+    }
+
+    fn do_chan_open_try_and_send_with_retry(&mut self) -> Result<(), ChannelError> {
+        retry_with_index(retry_strategy::default(), |_| {
+            self.do_chan_open_try_and_send()
+        })
+        .map_err(|err| {
+            error!("failed to open channel after {} retries", err);
+            ChannelError::Failed(format!(
+                "Failed to finish channel open try in {} iterations for {:?}",
+                retry_count(&err),
+                self
+            ))
+        })?;
+
+        Ok(())
+    }
+
+    /// Does a single step towards finalizing the channel open handshake.
+    /// Covers only a step of type either Ack or Confirm.
+    /// (Assumes that the channel open handshake was previously
+    /// initialized with Init & Try steps.)
+    ///
+    /// Returns `Ok` when both channel ends are in state `Open`.
+    /// Also returns `Ok` if the channel is undergoing a closing handshake.
+    ///
+    /// An `Err` can signal two cases:
+    ///     - the attempted handshake step has failed,
+    ///     - the attempted step may have finished successfully, but further
+    ///     steps are necessary to finalize the channel open handshake.
+    /// In both `Err` cases, there should be retry calling this method.
+    fn do_chan_open_ack_confirm_step(&self) -> Result<(), ChannelError> {
+        let src_channel_id = self
+            .src_channel_id()
+            .ok_or(ChannelError::MissingLocalChannelId)?;
+
+        let dst_channel_id = self
+            .dst_channel_id()
+            .ok_or(ChannelError::MissingCounterpartyChannelId)?;
+
+        // Continue loop if query error
+        let a_channel = self
+            .src_chain()
+            .query_channel(&self.src_port_id(), src_channel_id, Height::zero())
+            .map_err(|_| {
+                ChannelError::Failed(format!(
+                    "failed to query source chain {}",
+                    self.src_chain().id()
+                ))
+            })?;
+
+        let b_channel = self
+            .dst_chain()
+            .query_channel(&self.dst_port_id(), dst_channel_id, Height::zero())
+            .map_err(|_| {
+                ChannelError::Failed(format!(
+                    "failed to query destination chain {}",
+                    self.dst_chain().id()
+                ))
+            })?;
+
+        match (a_channel.state(), b_channel.state()) {
+            // Handle sending the Ack message to the source chain
+            (State::Init, State::TryOpen) | (State::TryOpen, State::TryOpen) => {
+                let event = self.flipped().build_chan_open_ack_and_send().map_err(|e| {
+                    error!("failed ChanAck {:?}: {}", self.a_side, e);
+                    e
+                })?;
+
+                info!(
+                    "done with ChanAck step {} => {:#?}\n",
+                    self.src_chain().id(),
+                    event
+                );
+                // One more step (confirm) left.
+                // Returning error signals that the caller should retry.
+                Err(ChannelError::PartialOpenHandshake(
+                    *a_channel.state(),
+                    *b_channel.state(),
+                ))
+            }
+
+            // Handle sending the Ack message to the destination chain
+            (State::TryOpen, State::Init) => {
+                let event = self.build_chan_open_ack_and_send().map_err(|e| {
+                    error!("failed ChanAck {:?}: {}", self.b_side, e);
+                    e
+                })?;
+
+                info!(
+                    "done with ChanAck step {} => {:#?}\n",
+                    self.dst_chain().id(),
+                    event
+                );
+                // One more step (confirm) left.
+                // Returning error signals that the caller should retry.
+                Err(ChannelError::PartialOpenHandshake(
+                    *a_channel.state(),
+                    *b_channel.state(),
+                ))
+            }
+
+            // Handle sending the Confirm message to the destination chain
+            (State::Open, State::TryOpen) => {
+                let event = self.build_chan_open_confirm_and_send().map_err(|e| {
+                    error!("failed ChanConfirm {:?}: {}", self.b_side, e);
+                    e
+                })?;
+
+                info!("done {} => {:#?}\n", self.dst_chain().id(), event);
+                Ok(())
+            }
+
+            // Send Confirm to the source chain
+            (State::TryOpen, State::Open) => {
+                let event = self
+                    .flipped()
+                    .build_chan_open_confirm_and_send()
+                    .map_err(|e| {
+                        error!("failed ChanConfirm {:?}: {}", self.a_side, e);
+                        e
+                    })?;
+
+                info!(
+                    "finalized channel open handshake {} => {:#?}\n",
+                    self.src_chain().id(),
+                    event
+                );
+                Ok(())
+            }
+            (State::Open, State::Open) => {
+                info!("channel handshake already finished for {:#?}\n", self);
+                Ok(())
+            }
+            // In all other conditions, return Ok, since the channel open handshake does not apply.
+            _ => Ok(()),
+        }
+    }
+
+    /// Takes a partially open channel and finalizes the open handshake protocol.
+    ///
+    /// Pre-condition: the channel identifiers are already established on both ends
+    ///   (i.e., the OpenInit and OpenTry steps have been fulfilled for this channel).
+    ///
+    /// Post-condition: the channel state is `Open` on both ends if successful.
+    fn do_chan_open_finalize_with_retry(&self) -> Result<(), ChannelError> {
+        retry_with_index(retry_strategy::default(), |_| {
+            self.do_chan_open_ack_confirm_step()
+        })
+        .map_err(|err| {
+            error!("failed to open channel after {} retries", err);
+            ChannelError::Failed(format!(
+                "Failed to finish channel handshake in {} iterations for {:?}",
+                retry_count(&err),
+                self
+            ))
+        })?;
+
+        Ok(())
+    }
+
     /// Executes the channel handshake protocol (ICS004)
     fn handshake(&mut self) -> Result<(), ChannelError> {
-        let done = '🥳';
-
-        let a_chain = self.src_chain().clone();
-        let b_chain = self.dst_chain().clone();
-
-        // Try chanOpenInit on a_chain
-        let mut counter = 0;
-        let mut init_success = false;
-        while counter < MAX_RETRIES {
-            counter += 1;
-            match self.flipped().build_chan_open_init_and_send() {
-                Err(e) => {
-                    error!("Failed ChanInit {:?}: {:?}", self.a_side, e);
-                    continue;
-                }
-                Ok(event) => {
-                    self.a_side.channel_id = Some(extract_channel_id(&event)?.clone());
-                    println!("{}  {} => {:#?}\n", done, a_chain.id(), event);
-                    init_success = true;
-                    break;
-                }
-            }
-        }
-
-        // Check that the channel was created on a_chain
-        if !init_success {
-            return Err(ChannelError::Failed(format!(
-                "Failed to finish channel open init in {} iterations for {:?}",
-                MAX_RETRIES, self
-            )));
-        };
-
-        // Try chanOpenTry on b_chain
-        counter = 0;
-        let mut try_success = false;
-        while counter < MAX_RETRIES {
-            counter += 1;
-            match self.build_chan_open_try_and_send() {
-                Err(e) => {
-                    error!("Failed ChanTry {:?}: {:?}", self.b_side, e);
-                    continue;
-                }
-                Ok(event) => {
-                    self.b_side.channel_id = Some(extract_channel_id(&event)?.clone());
-                    println!("{}  {} => {:#?}\n", done, b_chain.id(), event);
-                    try_success = true;
-                    break;
-                }
-            }
-        }
-
-        if !try_success {
-            return Err(ChannelError::Failed(format!(
-                "Failed to finish channel open try in {} iterations for {:?}",
-                MAX_RETRIES, self
-            )));
-        };
-
-        counter = 0;
-        while counter < MAX_RETRIES {
-            counter += 1;
-
-            let src_channel_id = self
-                .src_channel_id()
-                .ok_or(ChannelError::MissingLocalChannelId)?;
-            let dst_channel_id = self
-                .dst_channel_id()
-                .ok_or(ChannelError::MissingCounterpartyChannelId)?;
-            // Continue loop if query error
-            let a_channel =
-                a_chain.query_channel(&self.src_port_id(), &src_channel_id, Height::zero());
-            if a_channel.is_err() {
-                continue;
-            }
-            let b_channel =
-                b_chain.query_channel(&self.dst_port_id(), &dst_channel_id, Height::zero());
-            if b_channel.is_err() {
-                continue;
-            }
-
-            match (a_channel.unwrap().state(), b_channel.unwrap().state()) {
-                (State::Init, State::TryOpen) | (State::TryOpen, State::TryOpen) => {
-                    // Ack to a_chain
-                    match self.flipped().build_chan_open_ack_and_send() {
-                        Err(e) => error!("Failed ChanAck {:?}: {}", self.a_side, e),
-                        Ok(event) => println!("{}  {} => {:#?}\n", done, a_chain.id(), event),
-                    }
-                }
-                (State::Open, State::TryOpen) => {
-                    // Confirm to b_chain
-                    match self.build_chan_open_confirm_and_send() {
-                        Err(e) => error!("Failed ChanConfirm {:?}: {}", self.b_side, e),
-                        Ok(event) => println!("{}  {} => {:#?}\n", done, b_chain.id(), event),
-                    }
-                }
-                (State::TryOpen, State::Open) => {
-                    // Confirm to a_chain
-                    match self.flipped().build_chan_open_confirm_and_send() {
-                        Err(e) => error!("Failed ChanConfirm {:?}: {}", self.a_side, e),
-                        Ok(event) => println!("{}  {} => {:#?}\n", done, a_chain.id(), event),
-                    }
-                }
-                (State::Open, State::Open) => {
-                    println!(
-                        "{}  {}  {}  Channel handshake finished for {:#?}\n",
-                        done, done, done, self
-                    );
-                    return Ok(());
-                }
-                _ => {} // TODO channel close
-            }
-        }
-
-        Err(ChannelError::Failed(format!(
-            "Failed to finish channel handshake in {} iterations for {:?}",
-            MAX_RETRIES, self
-        )))
+        self.do_chan_open_init_and_send_with_retry()?;
+        self.do_chan_open_try_and_send_with_retry()?;
+        self.do_chan_open_finalize_with_retry()
     }
 
     pub fn counterparty_state(&self) -> Result<State, ChannelError> {