+++ relayer/src/light_client/mock.rs
@@ -1,8 +1,9 @@
+use ibc::ics02_client::misbehaviour::MisbehaviourEvidence;
 use tendermint_testgen::light_block::TmLightBlock;
 
 use ibc::ics02_client::client_state::AnyClientState;
 use ibc::ics02_client::events::UpdateClient;
-use ibc::ics02_client::misbehaviour::AnyMisbehaviour;
+use ibc::ics07_tendermint::header::Header as TmHeader;
 use ibc::ics24_host::identifier::ChainId;
 use ibc::mock::host::HostBlock;
 use ibc::Height;
@@ -10,6 +11,8 @@ use ibc::Height;
 use crate::chain::mock::MockChain;
 use crate::chain::Chain;
 use crate::error::Error;
+
+use super::Verified;
 
 /// A light client serving a mock chain.
 pub struct LightClient {
@@ -35,8 +38,11 @@ impl super::LightClient<MockChain> for LightClient {
         _trusted: Height,
         target: Height,
         _client_state: &AnyClientState,
-    ) -> Result<TmLightBlock, Error> {
-        Ok(self.light_block(target))
+    ) -> Result<Verified<TmLightBlock>, Error> {
+        Ok(Verified {
+            target: self.light_block(target),
+            supporting: Vec::new(),
+        })
     }
 
     fn fetch(&mut self, height: Height) -> Result<TmLightBlock, Error> {
@@ -47,7 +53,33 @@ impl super::LightClient<MockChain> for LightClient {
         &mut self,
         _update: UpdateClient,
         _client_state: &AnyClientState,
-    ) -> Result<Option<AnyMisbehaviour>, Error> {
+    ) -> Result<Option<MisbehaviourEvidence>, Error> {
         unimplemented!()
+    }
+
+    fn header_and_minimal_set(
+        &mut self,
+        trusted_height: Height,
+        target_height: Height,
+        client_state: &AnyClientState,
+    ) -> Result<Verified<TmHeader>, Error> {
+        let Verified { target, supporting } =
+            self.verify(trusted_height, target_height, client_state)?;
+
+        assert!(supporting.is_empty());
+
+        let succ_trusted = self.fetch(trusted_height.increment())?;
+
+        let target = TmHeader {
+            signed_header: target.signed_header,
+            validator_set: target.validators,
+            trusted_height,
+            trusted_validator_set: succ_trusted.validators,
+        };
+
+        Ok(Verified {
+            target,
+            supporting: Vec::new(),
+        })
     }
 }